Fix category pages by removing redirects and adding proper pages
